requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
python-multipart==0.0.9
sse-starlette==2.1.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Validation
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.3.0
base58==2.1.1

# Payments
stripe==10.12.0

# Solana (optional - install when connecting real chain)
# solana==0.34.3
# solders==0.21.0

# HTTP client
httpx==0.27.2

# Background tasks
celery[redis]==5.4.0
redis==5.0.8

# Email
resend==2.4.0

# Utils
python-dotenv==1.0.1
orjson==3.10.7
toml==0.10.2
loguru==0.7.2
sentry-sdk[fastapi]==2.14.0
